Moved the bound tests around. Added one for Aabb3
